# Inline project methods