Sometime these methods are not inline under `opt-level=z` compile option, forcing inlining could be beneficial for performance and binary size.