@taiki-e 2023-03-01
Thanks for the PR!

It is not clear that it is always reasonable to do this, since a large number of inline attributes will affect compile time (e.g., see https://github.com/rust-lang/hashbrown/pull/119)

Also, if you want to fully inline project_replace without LTO, the inline attribute may also be needed for [private types in pin-project-lite](https://github.com/taiki-e/pin-project-lite/blob/8a8b6e01a7de46ecd9ac1f7400d2b06e9fbf9cd7/src/lib.rs#L1531-L1575).

In any case, please provide specific data on what this will improve. (generated asm[^1], reproducible benchmarks, etc.)

[^1]: depend on the compiler version, but pin-project-lite is available on godbolt: https://godbolt.org/z/eTxGzb5E7