@EFanZh 2023-03-04
Here is my attempt at comparing the effect of inlining: <https://github.com/EFanZh/pin-project-lite/tree/compare-inlining>. (Sorry for not using godbolt, I couldn’t figure out how to patch `pin-project-lite` with a inlined version.)

Under `opt-level="z"`, here are the assembly code generated before and after inlining:

- Not inlined: <https://github.com/EFanZh/pin-project-lite/blob/compare-inlining/compare-original-oz.s>.
- Inlined: <https://github.com/EFanZh/pin-project-lite/blob/compare-inlining/compare-inline-oz.s>.

Also, I have tested the effects of `#[inline]` with `opt-level=3`, and the result is the same:

- Not inlined: <https://github.com/EFanZh/pin-project-lite/blob/compare-inlining/compare-original-o3.s>.
- Inlined: <https://github.com/EFanZh/pin-project-lite/blob/compare-inlining/compare-inlined-o3.s>.

These examples are handcrafted by me, maybe it’s better to test on a real project. But I don’t know any projects uses `pin-project-lite` extensively, is there any suggestions?