@EFanZh 2023-03-19
I tested this patch on [tokio-console](https://github.com/tokio-rs/console) with this branch: <https://github.com/EFanZh/console/tree/compare-inlining-pin-project-methods>. Here is the binary sizes comparison:

```text
original release     8145408
patched release      8138240

original release -z  6450688
patched release -z   6426112
```

Currently, I have only tested binary size changes. @taiki-e do you need other tests? For example: runtime performance benchmarks?