@taiki-e 2023-06-26
Sorry for the late reply. And thanks for the investigation. Assuming there is no impact on compile time, the effects of this all appear to be positive.

As for compile time, @nnethercote previously measured it by building async-std (https://github.com/taiki-e/pin-project-lite/pull/71). That said, I'm not sure if building the library is the right way to measure the impact of `#[inline]` on compile time. We may need to measure the compile time of binaries that use tokio, async-std, etc.