@EFanZh 2023-07-02
Again, I have tested the compile time of [tokio-console](https://github.com/tokio-rs/console) using [this script](https://github.com/EFanZh/console/blob/compare-inlining-pin-project-methods/benchmark-compiling.sh):

```bash
benchmark() {
    target_report=$1

    shift

    cargo clean
    cargo fetch "$@"
    cargo build --release --timings "$@"
    mv 'target/cargo-timings/cargo-timing.html' "$target_report"
}

patching_args=(
    --config 'patch.crates-io.pin-project-lite.git="https://github.com/EFanZh/pin-project-lite"' \
    --config 'patch.crates-io.pin-project-lite.rev="5ebf6921de62887572fa4e0477e725567de176df"'
)

for opt_level in '"z"' '3'; do
    for lto in 'off' 'thin' 'fat'; do
        for codegen_units in '16' '1'; do
            base_args=(
                --config "profile.release.opt-level=$opt_level"
                --config "profile.release.lto=\"$lto\""
                --config "profile.release.codegen-units=$codegen_units"
            )

            report_base_name="opt_level=${opt_level//\"/} lto=$lto codegen_units=$codegen_units"

            benchmark "$report_base_name - original.html" "${base_args[@]}"
            benchmark "$report_base_name - patched.html" "${base_args[@]}" "${patching_args[@]}"
        done
    done
done
```

Here is my test result:

| Configuration                         | Original | Patched |
| ------------------------------------- | -------- | ------- |
| opt_level=z lto=off codegen_units=16  | 51.7s    | 46.3s   |
| opt_level=z lto=off codegen_units=1   | 67.7s    | 67.6s   |
| opt_level=z lto=thin codegen_units=16 | 55.9s    | 55.9s   |
| opt_level=z lto=thin codegen_units=1  | 76.2s    | 69.4s   |
| opt_level=z lto=fat codegen_units=16  | 92.1s    | 91.7s   |
| opt_level=z lto=fat codegen_units=1   | 97.2s    | 97.3s   |
| opt_level=3 lto=off codegen_units=16  | 83.9s    | 83.4s   |
| opt_level=3 lto=off codegen_units=1   | 103.9s   | 103.6s  |
| opt_level=3 lto=thin codegen_units=16 | 68.6s    | 68.6s   |
| opt_level=3 lto=thin codegen_units=1  | 115.2s   | 115.6s  |
| opt_level=3 lto=fat codegen_units=16  | 140.2s   | 137.2s  |
| opt_level=3 lto=fat codegen_units=1   | 141.9s   | 142.5s  |

I didn’t see significant compile time change for this change.