@taiki-e 2023-07-02
Published in v0.2.10. Thanks @EFanZh.